refactor: rename -rest to -service
