+++ pom.xml
@@ -11,8 +11,8 @@
     <groupId>com.salaboy.conferences</groupId>
     <artifactId>conference-email-service</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>From Monolith to K8S :: Email Service (REST)</name>
-    <description>From Monolith to K8S :: Email Service (REST)</description>
+    <name>From Monolith to K8S :: Email Service</name>
+    <description>From Monolith to K8S :: Email Service</description>
 
     <properties>
         <java.version>17</java.version>
