+++ src/test/java/com/salaboy/conferences/email/EmailServiceTests.java
@@ -1,4 +1,4 @@
-package com.salaboy.conferences.email.rest;
+package com.salaboy.conferences.email;
